refactor: 통합 테스트에서 선택적으로 컨테이너를 붙이지 않으므로 하나의 파일로 통합
